baselayout: increase maxfiles, maxproc limits on mac
These limits are rather small by default.
Increasing the `maxproc' seems to be necessary to avoid:

  fork: resource temporarily unavailable

...when building QtWebKit with -j40, distcc and ccache.

Although I have not seen errors which can definitely be attributed to
`maxfiles' being too small, I increased it anyway, because the default
(soft) limit of 256 seems small enough to cause problems in the future.

Reviewed-by: Gareth Pethig

Change-Id: I507523a638d57bd7bf816273b7c278561d4e4feb