Fixed `hosts' module silently being ignored.
This module deploys a few entries into /etc/hosts.
It is supposed to be deployed on almost all machines.

It seems it has failed to deploy ever since the puppet repo was split
into private and public repos, and this wasn't noticed until now,
since all existing machines continued to work correctly.

We were attempting to have one module `include' another, which
appears to not work as expected.

Reviewed-by: Sergio Ahumada

Change-Id: I63118d5067f8b9eb15759165ec6f100fdce49c2e