network_test_server: fixed fragile management of danted services
The danted init script included in the installed package does not work
correctly when multiple danted instances are expected to be running, as
is the case for our network test server.  start-stop-daemon was matching
by usage of `--exec' and `--pidfile', but danted does not (and seemingly
has never) created a pid file, and `--exec' is insufficient to
differentiate the separate danted instances from each other.

This commit deploys the same init script for danted and
danted-authenticating and allows start-stop-daemon to manage the pid
file.

Reviewed-by: Kalle Lehtonen
Reviewed-by: Keith Isdale
Change-Id: Ie6c85b0de4519aaea9d3d2c55d1f7e835fc3cdd9