Fixed usage of sudo -i for setting HOME
According to the sudo man page, using the -i option should set HOME to
the value for the target user.  We depend on this behavior.
In practice, this can be overridden in /etc/sudoers, and it _is_
overridden in Mac OSX 10.7.

Use the -H option to explicitly opt-in to setting HOME everywhere we
need it.

Reviewed-by: Kalle Lehtonen
Change-Id: Ib27bafcbf0ef076d4245b8244e0cf00e1f42bea4