baselayout: temporarily disable core dumps on mac
Gathering core dumps on mac can take a while (~ a few minutes for a
Qt app on a mac mini).  This is problematic due to the fact
that qtbase master currently contains many crashing tests.

Disable the core dumps until the situation improves.  This means
that crashing tests will lack backtraces for a while.

Reviewed-by: Kalle Lehtonen
Change-Id: Ie0bd67372e2e30e4ff44d9422ee24b5509d4f23c