baselayout: re-enable core dumps on mac
This reverts commit 13577a34620683eb82a4c5230e9e2d1dadd109aa.

Core dumps were disabled because their collection is slow on mac,
and a very high number of crashing tests caused the mac CI
configurations to run for a long time (hours).

The offending tests have since been disabled and will not be re-enabled
until they are not crashing.  Therefore, it should be safe to enable
core dumps again.

Change-Id: Icd75c1946d76f6af9306a8f775cb90aec465688d
Reviewed-on: http://codereview.qt-project.org/5110
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>