pulseserver: double JVM heap size, move to 64-bit JVM
With our current workload, we are now capable of exhausting the JVM
heap.  Increase the size.

Note that this implies moving from a 32-bit to 64-bit JVM, as we cannot
increase the heap size in the 32-bit JVM without decreasing the thread
stack size or the thread count, both of which are undesirable.

Change-Id: I2b2de44ad3e9a99d21d2d4806894c1605272cad8
Reviewed-on: http://codereview.qt-project.org/5650
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>