pulseserver: let the pulse server have a predictable working directory
Previously, the pulse server's working directory would not be enforced
to any particular value.  It would inherit the working directory of, for
example, the ssh session used to run `/etc/init.d/pulseserver start'.

In practice this meant that having a working directory of /root was
possible, which was not readable by the pulseserver user account,
causing subtle behavior differences or problems in our post-build hook
scripts.

Change-Id: I89b80575b2581d871fee0ddadf9b41f546b9c8ff
Reviewed-on: http://codereview.qt-project.org/5905
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>