pulseserver: make publish_log_hook (very close to) atomic
We were writing directly to the log file, so it was possible for the log
file to be incomplete (e.g. if somebody downloaded it from the web server
while it was in the process of uploading).

Use a two stage process (write it to a file, then move it into the final
path) to make it more atomic.

Change-Id: I6f6822c792c10410ee6cf4f20b39ebc45b76c57c
Reviewed-by: Jyri Tahtela <jyri.tahtela@nokia.com>