pulseagent: run the agent inside a screen session to survive X crashes
We want to run the Pulse agent inside of a typical user desktop session,
since that is the primary environment targeted by Qt.
However, doing this means that the Pulse agent can be terminated
ungracefully if the X server crashes or if the user session exits
for any other reason.

Running inside of screen allows the Pulse agent to survive an X crash.
After the crash, the session is probably too damaged to give proper
results, but at least the build does not hang and the X server crash
is visible from the test log.

Change-Id: I911e7b1704f6588e1446960cdd917615bad58522
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>