Added new "smallfs" module; a small filesystem for testing out-of-space
Tests may check the QT_TEST_SMALL_FS environment variable.
This will be set to the mount point of a small (a few MB), world-writable
filesystem.  Tests may choose to fill up this filesystem, to check the
behavior when "no space left on device" errors occur.

Requested for testing of qtjsondb.

Change-Id: Ia74f5f8c63ef9f488ef442479c558b54e285a9fc
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>
Reviewed-by: Carlos Duclos <carlos.duclos@nokia.com>