Added workaround for bad fqdn values on Windows
Use the new --facts-from-reverse-dns switch to set the hostname, domain
and fqdn facts to values retrieved by a reverse DNS lookup before the
puppet run.

This is needed because we can't set Windows' local hostname to the true
FQDN without breaking some systems depending on the uniqueness of the
NetBIOS name, which is limited to 15 characters only.  15 characters is
insufficient for our FQDN naming scheme.

Change-Id: Ie8f4e89e106882116340abc039d212d23c486f4b
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>