nodes.pp: removed unnecessary references to Nokia
The pseudo-hostnames for the linux and mac test machines used
an address containing test.qt.nokia.com, but the setup is not tied to
Nokia. Replace the suffix with test.qt-project.org to make this clear.

The test.qt.nokia.com suffix was retained on a few one-off servers
which are unlikely to be usable outside of Nokia.

Change-Id: I7d04d2925e754738c0bd50415a49aad3e73f2d39
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>