Added new module, jenkins_slave.
jenkins_slave deploys a jenkins node on Windows or Linux.
The slave name is equal to the fqdn of the test machine and must be
pre-configured in the jenkins master before this module can be used.

The Windows implementation required some minor refactoring and the
addition of an implementation of the 'startup' defined type on Windows.

The installation of the Java virtual machine was split into a separate
module, since both Pulse and Jenkins depend on this.

Note that no nodes are yet assigned as Jenkins slaves.

Change-Id: I23e14564e2bd65fda607527107f4d7b6e7aea5e0
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>