jenkins_slave: don't set permissions on workspace in Windows
Allowing puppet to set the permissions on the workspace directory
causes .exe files compiled during the build/test process to lack
executable permissions.

When puppet controls the permissions on folders, it administers
the permissions always as "Special Permissions", and the other
(standard) permissions are unset.

The permissions are applied to folders only, not files within the
folders, and are not explicitly applied recursively. The 'owner'
and 'group' ownership permissions thus are not propagated to any
subdirectories under the top-level directory.

When MSVC links a new .exe file within a directory with these
permissions, the file ends up with read and write permissions to
Administrators only, and execute permissions are not set for any
user or group, making the file unable to be executed.

To avoid this problem, simply avoid managing the permissions with
puppet, since the default permissions are sufficient to provide
access to our test user account.

Note that managing permissions on files does not have this problem, so
these have been left as-is.

Change-Id: I3da23898c9386bc7dabf703195cffc80a5c6e5fa
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>