pulseserver: skip pulseserver-git for all OS other than Linux
This test was previously skipped on Windows.

Although it does pass on some OSX machines, others have problems because
the script assumes git is installed at /usr/bin/git (which is not true
when git is installed from macports).

It would be nice to fix it to also pass on Mac, but since the module is
not expected to be used anywhere other than Linux and any change to the
script involves risk, just skip it instead.

Change-Id: Icb8c97a68511a599c4194435bdb6070c579289d0
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>