baselayout: downgrade user creation message from fatal to warning
On Mac, we currently require the test user account to be created
manually, not via puppet. To ensure this is noticed, we deliberately
generate a message to the syslog informing about this, if the test user
account doesn't exist.

However, this was a fatal error ('fail'), which doesn't make sense - it
prevents other unrelated puppet modules from being executed, and it
prevents the module from being tested in --noop mode.

Make it into a warning instead.

Change-Id: Iebebc531ab3bf338c15a24e8673029e848bca0a1
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>