Replaced ubuntu_backports with apt_backports
The process for using a backports repository is the same for Debian and
Ubuntu, and theoretically may be the same for any OS using 'apt' for
package management; refactor it into an apt_backports module which
understands the correct default values for Debian and Ubuntu.

Note that the .conf files were renamed to .pref files. The
apt_preferences man page says that files in /etc/apt/preferences.d
should have no extension or a .pref extension only. In practice, the
.conf files appeared to work on Ubuntu but not Debian.

Change-Id: If0126d6bbb44a9efd16cf51310f3c7b9329382ae
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>