baselayout, jenkins_slave: fix startup() argument handling
Previously, the behavior was not consistent; on mac, the arguments would
be placed verbatim into a shell script, while on Linux, they would be
placed verbatim into a .desktop file. This meant that using shell
redirection such as '|', '2>&1' would work on mac, but would (sometimes)
not work on Linux (depending on the implementation - although the
.desktop file spec suggests otherwise, at least KDE on Ubuntu 10.04
appears to invoke the Exec key of a .desktop file through the shell,
while Gnome on later Ubuntu versions does not).

Adjust baselayout::startup() to take an array and quote it
appropriately, so the caller does not have to know if the string ends up
in a shell script or .desktop file.

Adjust some jenkins_slave startup items to explicitly invoke the shell
when attempting to use shell redirection.

Note the Windows implementation is unmodified and it is recommended to
keep using the single string (not array) $arguments on that platform.

Change-Id: I086115faf20c57c0f3f6c1555737d651cf44f278
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>