registry: fixed bad dependency setup sometimes leading to a cycle
We were defining a File for qtqa-reg.pl inside of the first created
Registry::Value. This File would implicitly belong to that
Registry::Value. Therefore, if the first created value depended on a
later created value, that later value would depend on the file which
depends on the first value, resulting in a cycle.

Move out shared resources and variables into the base 'registry' class.
This is probably good practice for future modules with a similar
structure.

Change-Id: I28ba756c7bb42de9346985fadc0b6a1a26333c24
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>