Move CI tester setup from nodes.pp into new top-level module
Extricate the setup for CI testers from nodes.pp into a new module.
The new ci_tester class takes relevant parameters for all included
classes. Default values are set for compatibility with the
current nodes.pp.

This is a step towards eliminating the need for private_nodes and the
other private overlays.

Change-Id: I0c0a5eefdb1e4033794cd4f9982c239d67746c9b
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>