Added an interactively configurable external node classifier
Prior to this change, the type of host to be managed was based on the
hostname, as per default puppet behavior.  This required adding various
patterns into a private overlay repository (private_nodes module), and
naming hosts according to those patterns.

The new setup uses a simple configuration file which can be
interactively configured at bootstrap time, hopefully making it much
easier to get started with this repository. This also reduces the usage
of the private overlay repository (which should eventually be
eliminated).

Change-Id: Ide4f1d9360a19b16b3dcaedf993951725932a011
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>
Reviewed-by: Toby Tomkins <tjtomkins@gmail.com>