jenkins_slave: fixed test attempting to make use of nonexistent user account
Some versions of puppet complain when referring to a nonexistent user
account or group in certain resources, even when running in no-op mode;
so, use the current user ($::id) as the test user, and a well-known group.

Since it is possible (at least on Windows) for $::id to contain characters
with a special meaning when used in regular expressions, make sure to
quote it.

Change-Id: I0ad1a16aec8bcd31dc3361e883d8b3e9c40bc562
Reviewed-by: Toby Tomkins <tjtomkins@gmail.com>