Add support for top-level puppet variables to nodecfg.
In the past, private-overlay has been used to configure correct
values for the global variables. In order to get rid of private-
overlay, support for configuring the needed top-level puppet
variables with external node configuration (ENC) i.e. nodecfg.pl
was added.

Since qtgitreadonly_local is not used anymore, it was removed
from global configuration. In addition the whole private_nodes
module was removed from this repo since it is assumed that
the global variables are initialized with ENC.

Change-Id: I4e9a4f2c048e7020fa45e08c3710730a8feb628d
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>