Fixed sudo setup on OSX 10.6; use sudo from macports
OSX 10.6 ships a version of sudo too old to understand the '#includedir'
directive; this meant the passwordless reboot setup for the qt user did
not work.

Deploy a newer version of sudo from macports instead, allowing us to
manage sudoers entries the same way on OSX 10.6 as we do on OSX 10.7 and
Linux; fixes passwordless reboot.

Change-Id: I70d97eb708e399c5e50d534aa148e2cc06993c89
Reviewed-by: Toby Tomkins <tjtomkins@gmail.com>