jenkins_slave: retry online-node a few times at startup
When a test machine boots, it might happen that the
jenkins_slave::register_online startup item is launched before the
machine's network interface has been brought up, or the Jenkins server
might be temporarily unreachable for some other reason.

To be robust in this case, we should retry the online-node command a few
times; otherwise the slave may remain in offline state for longer than
necessary.

Change-Id: I11694f0f49c6562f15981e6891613ffc388b2965
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>