jenkins_slave: fixed 'group' on jenkins files and workdir
A mistake was made in the default $jenkins_slave::group parameter; it
was set to $baselayout::group, which doesn't exist.
$baselayout::testgroup is the correct variable.

This resulted in no particular group being enforced on any of the files;
usually not a problem, but on some mac systems, the workdir may be owned
by 'wheel', which causes a failure in the tst_qfileinfo autotest.

Change-Id: I08895ca6013162e1aa2725abe715554db93eb4c4
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>