jenkins_slave: reboot to recover from catastrophic failure
If the Jenkins slave was running successfully for some time, but then
expectedly exited, previously we would warn and restart it.

However, this can cause undesirable behavior in some situations: for
example, if the Jenkins slave was killed because the host has run out of
memory, we would re-run the slave and allow further builds to fail on
the same host.

If the Jenkins slave was terminated during a build, it might also happen
that the next build starts while processes remain from a previous build.

Change it to reboot whenever the Jenkins slave unexpectedly terminates.

Change-Id: If07d197813bcf828ee9925b5a4f4a1ba83ba7b09
Reviewed-by: Toby Tomkins <tjtomkins@gmail.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>