Remove deprecated hostnames from nodes.pp
These nodes are from now on configured with External Node Classifier (ENC).

We can very likely remove also other nodes listed in this file, but before
doing that I will check our plans for using old XServer and Solaris machines.
Even it is  very unlikely that we will ever add Solaris or other commercial
UNIX to 4.8 CI, it might be that these configurations are useful for release
creation and release testing system.

Change-Id: Iaf6e73123a23858fc5bd9346ab8737c6a0730802
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Rohan McGovern [inactive] <rohan.mcgovern@nokia.com>
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>