Make jom installation on Windows machines controlled by Puppet.
There has been a few build failures recently in some CI projects
which are suspected to be caused by Jom bug. This puppet configuration
makes it easy to upgrade/downgrade jom in Windows CI testers when
needed.

Note that this configuration does not add jom to path, but it needs
to be done manually or preferably using Jenkins configuration.

Change-Id: I75ab3029fba601c286c562890893214accf5b271
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Simo Fält <simo.falt@digia.com>
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>