Make sure macports is updated before using it.
Add exec to run macports selfupdate & sync so that changes to
macports.conf and sources.conf are applied.

Use puppet 'require' instead of 'include' for macports where possible
to make sure that macports stuff is applied before macports are used.

In case of baselayout, 'require' cannot be used because it will end-up
to cyclic dependency. So lets use explicit 'require' for 'sudo' package
only  in that case.

Change-Id: I05864e0532514f8f7bc51ca5afd38eacb64a474c
Reviewed-by: Rohan McGovern <rohan@mcgovern.id.au>