Fix perl local::lib installation on Mac.
Download tarball to correct directory. I'm not quite sure why this has
been ever working or did the tempdir behaviour changed when we switched
from build-in perl to MacPorts perl. However with MacPorts perl the tarball
was downloaded to 'tempdir' under 'tempdir', this because there is
'chdir( $tempdir )' command before fetch. As an result of this extracting
tarball failed because archive was not found.

In addition 10.7. and 10.8 machines complained about '/bin/sh -c' command
as follows: '/bin/sh: -c: option requires an argument'. This was fixed by
removing the newlines from the command.

The perl script also have dependency to p5-libwww-perl, so make sure it is
installed.

Change-Id: I57d71d40c0720f8925b44b19cb0c23ba5c406fd5
Reviewed-by: Rohan McGovern <rohan@mcgovern.id.au>