Fix MSI uninstaller copying under installation folder.
In case that SW with MSI installer is installed manually before
puppet is executed, the current approach gives error from file
resource because $real_install_package does not exist.
This is the case in our current Win7 CI slaces, so lets change
the logic so that copying is done as part of install exec resource.

Change-Id: I9d4837fa5510f7500a6ebf43d0d97fa1f93d75c1
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>