Puppet configuration to start Windows 8 in desktop mode at boot.
Windows 8 support in Qt is currently limited to desktop mode, so
let's ensure that OS is launched to destop mode when rebooted.

Without this config some Qt auto test (for example tst_qmenu) fail,
because the menu will not get focus on top of Metro style
"start screen" but only gets the focus in desktop mode.

Task-number: QTBUG-28611

Change-Id: I3053bde3c56f3c4c59a32be7a33014d094923bf7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>