Create a timeout for sync_and_run's lock file
If the lock file is over 90 minutes (configurable in sync_and_run.pl) old,
try to kill the process whose pid was found in the lock file. After that
try to remove the lock file and resume program.
In the case of a computer shutdown, perl processes are killed by OS
and lock file was left in place. This now enables Puppet to resume.
This also prevents Puppet runs that take over 90 minutes.

Nothing yet prevents a shutdown from still killing a Puppet run.

Change-Id: Ifec7bfa5a710ad5175d9fc2d867537b32dc19bf7
Reviewed-by: Heikki Halmet <heikki.halmet@digia.com>