Qt metrics v0.17: Metrics box description mechanism
Added descriptions to Project and Autotest dashboards (the "i" icon).
Descriptions are content (filtering level) specific like the content
in the metrics boxes themselves. This uses a new mechanism to include
description html files as new small windows.

Replaced the "i" icon for failure categories in Autotest dasbboard
(that showed the description when mouse moved over it) with a text
link to open a description window. This was because a third "i" icon
in the middle of the screen caught the eye and therefore bad for
usability. Link text style changed for usability and readability.

Change-Id: I3ca4ee5f716197f4a915173015a7b145989b8e40
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>