Qt metrics v1.1.1: Improvement to Databse status box
Changed the order how the boxes are loaded on initial page load.
It is now: Database status box -> Filters -> Metrics boxes (the
Database status box was previously the last one).
This is to indicate the possible database rebuild progress instead
of of welcome text when loading the page first time during rebuild.

Change-Id: I4077ce942dce56f49db81db741ec0c0d1e6247e9
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>