Qt metrics v1.3: Result history to Autotest dashboard
Last 20 (or as configured) Project Configurations builds checked and
the Autotest result printed for each based on the failure category.
Link to each build log file available. Timescale filtering shown in
the table. The notation description is available to the user as a
message window link.
(This change replaces the first one that was abandoned due to merge
problems)

Change-Id: I06ba2197126d95f2fc0b9793efd4bdaf7a57fb87
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>