Qt metrics v1.5.3: Ajax and metrics box structure refactoring
Some refactoring is needed to the Ajax module and its calls and
responses to make it easier to add new metrics pages and to avoid
duplicate code. This is pure implementation change and is not
visible to the user.
Changes include following:
The CI specific variable in metricsdefinitions.php was renamed to
be generic so that the same file structure can be used for new pages.
The multiple fixed filter parameters in get method changed to one
filter string so that the same interface can be used for different
needs.
The order of execution in initial page loading moved to one file
(metricspageci.php) to make it easier to adjust for different needs.
A common purpose function getLocalTime moved from ci/functions.php
to new file commonfunctions.php.

Change-Id: I899b8b7432f202ffbb67687b79ef9b8310a164bd
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>