Qt metrics v1.6: RTA metrics page created
New metrics page created for RTA (Release Test Automation). Data is read
from XML files located in a certain directory. The page includes initially
a list of RTA failures, and functionality to filter the results by test
type, license type and platform.
The RTA metrics page is not yet set visible in the main menu (that will be
done later when all the needed views and functionality have been
implemented).
Note: This RTA page will not be public, but to be visible in Digia
environment only.

Change-Id: I705a8083cdf4310abc94cb4cd6f40c48a661b142
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>