Qt metrics v1.8: RTA comparison view
RTA comparison view added to the RTA history metrics box as level 2 view.
The XML reading function for RTA failures moved to functions.php to be
able to use it for several metrics boxes. Common RTA definitions moved
to definitions.php to make them available for any module and to avoid
multiple definitions.

Change-Id: Ib90afe10da3fc450c234ea8f747d809010ac3593
Reviewed-by: Tarja Sundqvist <tarja.sundqvist@digia.com>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>