Qt metrics v2.0: Qt metrics main menu
Main menu item selection and identification of the active page
implemented. Current menu items are CI metrics and RTA metrics
(latter not publicly available). The server and client check
functions added to be able to identify public and internal server,
and to be able to tailor functionality based on that. The connection
definitions arranged to common part and server specific part.

Change-Id: Ic468308764cf03e91259bda9115428b50e34836c
Reviewed-by: Tarja Sundqvist <tarja.sundqvist@digia.com>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>