Qt Metrics v2.1: Performance optimization
Changed both the CI project and autotest dashboards to utilize the
new "latest" database tables that include data only for the latest
builds per each project. Reading data from these tables is much
quicker than from the complete tables because of the size of the
tables and because there is no need to check the latest build number
in database searches.

Added functionality to the metrics engine to repeat calling one metrics
box several times per one update. Changed both the CI dashboards to
use the two rounds to load the data in levels 1-3. The first round
loads the latest build data and the second round loads also the all
builds data. This will ensure that the first round data will be visible
very quickly while the second round may take a bit longer. Made also
the required modifications to the RTA metrics because of the changes
to the metrics engine.

Changed the list cut mode in CI project dashboard to the short list
(with text "show full list") only on the first page load like with
the autotest dashboard.

Change-Id: Ic4ace5a53521a141d5602c2cdd61df7725c8fd64
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>