Qt Metrics v2.3: Select any build
Added functionality to select a build for a project or its configuration.
Basic approach is to use the timescale filter as a master switch: if not
set then show the latest/selected build data, and if set then show
latest/selected build data on the left-hand side and all builds data
(since the selected date) on the right-hand side. Added color indication
to the selected build and timescale to visualize the selections.

Included a loading indication as a pop-up message when the data loading
operation lasts longer than three seconds (the delay is configurable).

Change-Id: I65e2a7b5ef4c128498eb402da47a3f0e593c40f9
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>