Qt Metrics v2.4: Performance improvement to autotest
Performance improvement to autotest dashboard test result data. Read
the number of run autotests from the database instead of test result
zip files for better performance. New tables all_test and all_test_latest
used. Added showing the detailed test result data also when project is not
filtered, with a function to show or hide the list to optimize the
performance if the user does not need this data.

Change-Id: Id43115e019e3f67293f459d2d31197ec77990631
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>