Qt Metrics v2.5: Split project filter
Split the current project filter to project name and branch.
The filters and the links in the dashboards provided the same
functionality and were therefore unnecessary duplicate to each
other. Branch name is identified at the end of the full project
name (between the last two "_" i.e. it may not contain "_"
character), and the project name is identified to be the string
till the branch name (it may contain "_" characters).

The new project name and branch filters affect to level 1 in Project
dashboard and to level 1 and 2 in Autotest dashboard. Autotest
dashboard level 1 counting changed to display the number of
configurations in each project i.e. same configuration in e.g.
three projects will result to three (earlier resulted to one).
New failure summary table added to Autotest dashboard level 2 to
visualize the numbers shown on level 1 for the latest or selected build.

Change-Id: I6134f650d3df175c3655e31379d04cad2d869da3
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>