Qt Metrics v2.6: Platform filter
Added the Platform filter. I uses a predefined list of searches
(in ci/definitions.php) to the configurations i.e. lists the
Projects built for the selected Configurations, and Autotests
failed in those Configurations. The Configuration filter works
in the same way but includes only one Configuration, and it
overwrites the Platform filter as such.

The filters were rearranged inside the filter box.

Change-Id: I45d585c99e1ec02acb8945d1cf102c20855ecfec
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>