Force jenkins win slave to reboot when slave exits
To fix issues where jenkins win slave gives an exception right
after starting a job, we will reboot the node when there is some
problems with the slave's connection.

The root cause for the original problem seems to be unreliable
network connection, where master never gets the information
that the node went down. After reboot when node is trying to
re-connect, master rejects the request, while it thinks
the node is already connected.

Task-number: QTQAINFRA-846
Change-Id: I6edc4d9ab57f63121e5148f3449b7fa3dea021d5
Reviewed-by: Sergio Ahumada <sahumada@blackberry.com>