Qt Metrics 2 (v0.7): Overview and Project pages
Implemented the overview page to show status of the
latest Qt5 state build for each branch including
the build result itself and the testset results in
the Qt5 state builds (testset results are shown by
testset project).
Implemented the build project (Qt5 state in practice)
and testset project pages, the first to show the
build results and the latter to show the testset
results. Both are linked from the overview page,
and the testset project page also from the new
search box on the home page.

Change-Id: Iaa97597b63e05af426e6749af2b2cbadbb492eb7
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@theqtcompany.com>