Qt Metrics 2 (v0.31): Blacklisted passed testfunctions page
New page to list blacklisted passed test functions
since a specific date to be able to point out the
ones where the blacklisted tag could maybe be removed.
The date scope can be configured in the ini file.
The query duration is about 50 seconds with the
scope of last five days.

Change-Id: I21c37182c282d68838ec432c06f4d1c428312d48
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@theqtcompany.com>