Qt Metrics 2 (v0.35): Top testset/testfunction duration
New pages to list top testset durations and testfunction
durations (for one testset) since a specific date including
the runs that take longer than a specific limit.
The testset duration list can be opened from the button
on home page, and the testfunction list from a button
on the testset page.

Change-Id: I19d2fe1a8cc06f68cb78c7672148687b0fdff681
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@theqtcompany.com>