Fix policy failing sometimes to start danted
The policy starts several processes, of two types:

/usr/sbin/danted -f /etc/danted.conf
/usr/sbin/danted -f /etc/danted-authenticating.conf

If it happened that danted-authenticating was started first, then puppet
found a match for "danted" in the process table and assumed "danted"
was running, falsely.

By adding a specific "pattern" to search for in the process table,
the two services can now start in any order.

Additionally, when rebooting the system, stopping "danted" left
the PIDfile in /var/run. After reboot, the stale PIDfile happened
sometimes to match an existing process, and the "start-stop-daemon"
program as invoked by the initscript would refuse to start the
program, complaining it's already running.

Now we clean up the PIDfile after stopping the service,
to make sure this will never happen again.

Finally we moved the PIDfiles that start-stop-server writes
to their own directory, because danted always writes
to danted.pid, even when started multiple times with
different configuration files! Like that, danted used to
overwrite the PIDfiles of start-stop-server, and the
initscript got confused.

Last, remove the quiet flag from start-stop-server, so that we
get a message like "process already running" in the log,
if a failure occurs again.

Change-Id: I2cd996eb5ef2f14e571785221319f0c98e936fde
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@qt.io>