Add an updated version of the httpcachetest_expires304 cgi script
The cgi script was comparing the value of "if modified since" with the
date it advertises that the page will _expire_. And only then will it
return a "304 Not Modified". This is obviously not how the server should
respond.

The change to the code in QtNetwork itself will not be in e.g. 5.12 so
we need to copy the script and direct the fixed versions to point there.

Change-Id: I869d2e31fdd2a799fdc7083d2a73d15898d26429
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>